Move the main validate function to entrypoint.py and change validate.py to validator.py
