+++ pyshacl/validator.py
@@ -1,13 +1,9 @@
 # -*- coding: utf-8 -*-
 #
 import logging
-import os
 import sys
-from functools import wraps
-from io import BufferedIOBase, TextIOBase
 from os import getenv, path
-from sys import stderr
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
 import rdflib
 from rdflib import BNode, Literal, URIRef
@@ -19,10 +15,9 @@ from .consts import (
     SH_ValidationReport,
     env_truths,
 )
-from .errors import ReportableRuntimeError, ValidationFailure
+from .errors import ReportableRuntimeError
 from .extras import check_extra_installed
 from .functions import apply_functions, gather_functions, unapply_functions
-from .monkey import apply_patches, rdflib_bool_patch, rdflib_bool_unpatch
 from .pytypes import GraphLike, SHACLExecutor
 from .rdfutil import (
     add_baked_in,
@@ -30,27 +25,59 @@ from .rdfutil import (
     clone_graph,
     inoculate,
     inoculate_dataset,
-    load_from_source,
     mix_datasets,
     mix_graphs,
 )
 from .rules import apply_rules, gather_rules
 from .shapes_graph import ShapesGraph
 from .target import apply_target_types, gather_target_types
-from .validator_conformance import check_dash_result
 
 USE_FULL_MIXIN = getenv("PYSHACL_USE_FULL_MIXIN") in env_truths
 
-log_handler = logging.StreamHandler(stderr)
-log = logging.getLogger(__name__)
-for h in log.handlers:
-    log.removeHandler(h)  # pragma:no cover
-log.addHandler(log_handler)
-log.setLevel(logging.INFO)
-log_handler.setLevel(logging.INFO)
-
 
 class Validator(object):
+    def __init__(
+        self,
+        data_graph: GraphLike,
+        *args,
+        shacl_graph: Optional[GraphLike] = None,
+        ont_graph: Optional[GraphLike] = None,
+        options: Optional[Dict[str, Any]] = None,
+        **kwargs,
+    ):
+        options = options or {}
+        self._load_default_options(options)
+        self.options = options  # type: dict
+        self.logger = options['logger']  # type: logging.Logger
+        self.debug = options['debug']
+        self.pre_inferenced = kwargs.pop('pre_inferenced', False)
+        self.inplace = options['inplace']
+        if not isinstance(data_graph, rdflib.Graph):
+            raise RuntimeError("data_graph must be a rdflib Graph object")
+        self.data_graph = data_graph  # type: GraphLike
+        self._target_graph = None
+        self.ont_graph = ont_graph  # type: Optional[GraphLike]
+        self.data_graph_is_multigraph = isinstance(self.data_graph, (rdflib.Dataset, rdflib.ConjunctiveGraph))
+        if self.ont_graph is not None and isinstance(self.ont_graph, (rdflib.Dataset, rdflib.ConjunctiveGraph)):
+            self.ont_graph.default_union = True
+        if self.ont_graph is not None and options['sparql_mode']:
+            raise ReportableRuntimeError("Cannot use SPARQL Remote Graph Mode with extra Ontology Graph inoculation.")
+        if shacl_graph is None:
+            if options['sparql_mode']:
+                raise ReportableRuntimeError(
+                    "SHACL Shapes Graph must be a separate local graph or file when in SPARQL Remote Graph Mode."
+                )
+            shacl_graph = clone_graph(data_graph, identifier='shacl')
+        assert isinstance(shacl_graph, rdflib.Graph), "shacl_graph must be a rdflib Graph object"
+        self.shacl_graph = ShapesGraph(shacl_graph, self.debug, self.logger)  # type: ShapesGraph
+
+        if options['use_js']:
+            if options['sparql_mode']:
+                raise ReportableRuntimeError("Cannot use SHACL-JS in SPARQL Remote Graph Mode.")
+            is_js_installed = check_extra_installed('js')
+            if is_js_installed:
+                self.shacl_graph.enable_js()
+
     @classmethod
     def _load_default_options(cls, options_dict: dict):
         options_dict.setdefault('debug', False)
@@ -162,48 +189,6 @@ class Validator(object):
                             cloned_nodes[(source, _id)] = o = URIRef(_id)
                 vg.add((s, p, o))
         return vg, v_text
-
-    def __init__(
-        self,
-        data_graph: GraphLike,
-        *args,
-        shacl_graph: Optional[GraphLike] = None,
-        ont_graph: Optional[GraphLike] = None,
-        options: Optional[dict] = None,
-        **kwargs,
-    ):
-        options = options or {}
-        self._load_default_options(options)
-        self.options = options  # type: dict
-        self.logger = options['logger']  # type: logging.Logger
-        self.debug = options['debug']
-        self.pre_inferenced = kwargs.pop('pre_inferenced', False)
-        self.inplace = options['inplace']
-        if not isinstance(data_graph, rdflib.Graph):
-            raise RuntimeError("data_graph must be a rdflib Graph object")
-        self.data_graph = data_graph  # type: GraphLike
-        self._target_graph = None
-        self.ont_graph = ont_graph  # type: Optional[GraphLike]
-        self.data_graph_is_multigraph = isinstance(self.data_graph, (rdflib.Dataset, rdflib.ConjunctiveGraph))
-        if self.ont_graph is not None and isinstance(self.ont_graph, (rdflib.Dataset, rdflib.ConjunctiveGraph)):
-            self.ont_graph.default_union = True
-        if self.ont_graph is not None and options['sparql_mode']:
-            raise ReportableRuntimeError("Cannot use SPARQL Remote Graph Mode with extra Ontology Graph inoculation.")
-        if shacl_graph is None:
-            if options['sparql_mode']:
-                raise ReportableRuntimeError(
-                    "SHACL Shapes Graph must be a separate local graph or file when in SPARQL Remote Graph Mode."
-                )
-            shacl_graph = clone_graph(data_graph, identifier='shacl')
-        assert isinstance(shacl_graph, rdflib.Graph), "shacl_graph must be a rdflib Graph object"
-        self.shacl_graph = ShapesGraph(shacl_graph, self.debug, self.logger)  # type: ShapesGraph
-
-        if options['use_js']:
-            if options['sparql_mode']:
-                raise ReportableRuntimeError("Cannot use SHACL-JS in SPARQL Remote Graph Mode.")
-            is_js_installed = check_extra_installed('js')
-            if is_js_installed:
-                self.shacl_graph.enable_js()
 
     @property
     def target_graph(self):
@@ -412,211 +397,3 @@ def assign_baked_in():
     add_baked_in("http://datashapes.org/schema", schema_file)
     add_baked_in("https://datashapes.org/schema", schema_file)
     add_baked_in("http://datashapes.org/schema.ttl", schema_file)
-
-
-def with_metashacl_shacl_graph_cache(f):
-    # noinspection PyPep8Naming
-    EMPTY = object()
-
-    @wraps(f)
-    def wrapped(*args, **kwargs):
-        graph_cache = getattr(wrapped, "graph_cache", None)
-        assert graph_cache is not None
-        if graph_cache is EMPTY:
-            import pickle
-
-            if getattr(sys, 'frozen', False):
-                # runs in a pyinstaller bundle
-                here_dir = sys._MEIPASS
-            else:
-                here_dir = path.dirname(__file__)
-            pickle_file = path.join(here_dir, "assets", "shacl-shacl.pickle")
-            with open(pickle_file, 'rb') as shacl_pickle:
-                u = pickle.Unpickler(shacl_pickle, fix_imports=False)
-                shacl_shacl_store, identifier = u.load()
-            shacl_shacl_graph = rdflib.Graph(store=shacl_shacl_store, identifier=identifier)
-            setattr(wrapped, "graph_cache", shacl_shacl_graph)
-        return f(*args, **kwargs)
-
-    setattr(wrapped, "graph_cache", EMPTY)
-    return wrapped
-
-
-@with_metashacl_shacl_graph_cache
-def meta_validate(shacl_graph: Union[GraphLike, str], inference: Optional[str] = 'rdfs', **kwargs):
-    shacl_shacl_graph = meta_validate.graph_cache
-    shacl_graph = load_from_source(shacl_graph, rdf_format=kwargs.pop('shacl_graph_format', None), multigraph=True)
-    _ = kwargs.pop('meta_shacl', None)
-    return validate(shacl_graph, shacl_graph=shacl_shacl_graph, inference=inference, **kwargs)
-
-
-def validate(
-    data_graph: Union[GraphLike, BufferedIOBase, TextIOBase, str, bytes],
-    *args,
-    shacl_graph: Optional[Union[GraphLike, BufferedIOBase, TextIOBase, str, bytes]] = None,
-    ont_graph: Optional[Union[GraphLike, BufferedIOBase, TextIOBase, str, bytes]] = None,
-    advanced: Optional[bool] = False,
-    inference: Optional[str] = None,
-    inplace: Optional[bool] = False,
-    abort_on_first: Optional[bool] = False,
-    allow_infos: Optional[bool] = False,
-    allow_warnings: Optional[bool] = False,
-    max_validation_depth: Optional[int] = None,
-    sparql_mode: Optional[bool] = False,
-    focus_nodes: Optional[List[Union[str, URIRef]]] = None,
-    use_shapes: Optional[List[Union[str, URIRef]]] = None,
-    **kwargs,
-):
-    """
-    :param data_graph: rdflib.Graph or file path or web url of the data to validate
-    :type data_graph: rdflib.Graph | str | bytes
-    :param args:
-    :type args: list
-    :param shacl_graph: rdflib.Graph or file path or web url of the SHACL Shapes graph to use to
-    validate the data graph
-    :type shacl_graph: rdflib.Graph | str | bytes
-    :param ont_graph: rdflib.Graph or file path or web url of an extra ontology document to mix into the data graph
-    :type ont_graph: rdflib.Graph | str | bytes
-    :param advanced: Enable advanced SHACL features, default=False
-    :type advanced: bool | None
-    :param inference: One of "rdfs", "owlrl", "both", "none", or None
-    :type inference: str | None
-    :param inplace: If this is enabled, do not clone the datagraph, manipulate it in-place
-    :type inplace: bool
-    :param abort_on_first: Stop evaluating constraints after first violation is found
-    :type abort_on_first: bool | None
-    :param allow_infos: Shapes marked with severity of sh:Info will not cause result to be invalid.
-    :type allow_infos: bool | None
-    :param allow_warnings: Shapes marked with severity of sh:Warning or sh:Info will not cause result to be invalid.
-    :type allow_warnings: bool | None
-    :param max_validation_depth: The maximum number of SHACL shapes "deep" that the validator can go before reaching an "endpoint" constraint.
-    :type max_validation_depth: int | None
-    :param sparql_mode: Treat the DataGraph as a SPARQL endpoint, validate the graph at the SPARQL endpoint.
-    :type sparql_mode: bool | None
-    :param focus_nodes: A list of IRIs to validate only those nodes.
-    :type focus_nodes: list | None
-    :param kwargs:
-    :return:
-    """
-
-    do_debug = kwargs.get('debug', False)
-    if do_debug:
-        log_handler.setLevel(logging.DEBUG)
-        log.setLevel(logging.DEBUG)
-    apply_patches()
-    assign_baked_in()
-    do_check_dash_result = kwargs.pop('check_dash_result', False)  # type: bool
-    if kwargs.get('meta_shacl', False):
-        to_meta_val = shacl_graph or data_graph
-        conforms, v_r, v_t = meta_validate(to_meta_val, inference=inference, **kwargs)
-        if not conforms:
-            msg = f"SHACL File does not validate against the SHACL Shapes SHACL (MetaSHACL) file.\n{v_t}"
-            log.error(msg)
-            raise ReportableRuntimeError(msg)
-    do_owl_imports = kwargs.pop('do_owl_imports', False)
-    data_graph_format = kwargs.pop('data_graph_format', None)
-
-    if isinstance(data_graph, (str, bytes, BufferedIOBase, TextIOBase)):
-        # DataGraph is passed in as Text. It is not an rdflib.Graph
-        # That means we load it into an ephemeral graph at runtime
-        # that means we don't need to make a copy to prevent polluting it.
-        ephemeral = True
-    else:
-        ephemeral = False
-    use_js = kwargs.pop('js', None)
-    if sparql_mode:
-        if use_js:
-            raise ReportableRuntimeError("Cannot use SHACL-JS in SPARQL Remote Graph Mode.")
-        if inplace:
-            raise ReportableRuntimeError("Cannot use inplace mode in SPARQL Remote Graph Mode.")
-        if ont_graph is not None:
-            raise ReportableRuntimeError("Cannot use SPARQL Remote Graph Mode with extra Ontology Graph inoculation.")
-        if isinstance(data_graph, bytes):
-            data_graph = data_graph.decode('utf-8')
-        else:
-            data_graph = data_graph
-        ephemeral = False
-        inplace = True
-    if (
-        sparql_mode
-        and isinstance(data_graph, str)
-        and (data_graph.lower().startswith("http:") or data_graph.lower().startswith("https:"))
-    ):
-        from rdflib.plugins.stores.sparqlstore import SPARQLStore
-
-        query_endpoint: str = data_graph
-        username = os.getenv("PYSHACL_SPARQL_USERNAME", "")
-        method = os.getenv("PYSHACL_SPARQL_METHOD", "GET")
-        auth: Optional[Tuple[str, str]]
-        if username:
-            password: str = os.getenv("PYSHACL_SPARQL_PASSWORD", "")
-            auth = (username, password)
-        else:
-            auth = None
-        store = SPARQLStore(query_endpoint=query_endpoint, auth=auth, method=method)
-        loaded_dg = rdflib.Dataset(store=store, default_union=True)
-    else:
-        # force no owl imports on data_graph
-        loaded_dg = load_from_source(
-            data_graph, rdf_format=data_graph_format, multigraph=True, do_owl_imports=False, logger=log
-        )
-    ont_graph_format = kwargs.pop('ont_graph_format', None)
-    if ont_graph is not None:
-        loaded_og = load_from_source(
-            ont_graph, rdf_format=ont_graph_format, multigraph=True, do_owl_imports=do_owl_imports, logger=log
-        )
-    else:
-        loaded_og = None
-    shacl_graph_format = kwargs.pop('shacl_graph_format', None)
-    if shacl_graph is not None:
-        rdflib_bool_patch()
-        loaded_sg = load_from_source(
-            shacl_graph, rdf_format=shacl_graph_format, multigraph=True, do_owl_imports=do_owl_imports, logger=log
-        )
-        rdflib_bool_unpatch()
-    else:
-        loaded_sg = None
-    iterate_rules = kwargs.pop('iterate_rules', False)
-    if "abort_on_error" in kwargs:
-        log.warning("Usage of abort_on_error is deprecated. Use abort_on_first instead.")
-        ae = kwargs.pop("abort_on_error")
-        abort_on_first = bool(abort_on_first) or bool(ae)
-    validator_options_dict = {
-        'debug': do_debug or False,
-        'inference': inference,
-        'inplace': inplace or ephemeral,
-        'abort_on_first': abort_on_first,
-        'allow_infos': allow_infos,
-        'allow_warnings': allow_warnings,
-        'advanced': advanced,
-        'iterate_rules': iterate_rules,
-        'use_js': use_js,
-        'sparql_mode': sparql_mode,
-        'logger': log,
-        'focus_nodes': focus_nodes,
-        'use_shapes': use_shapes,
-    }
-    if max_validation_depth is not None:
-        validator_options_dict['max_validation_depth'] = max_validation_depth
-    validator = None
-    try:
-        validator = Validator(
-            loaded_dg,
-            shacl_graph=loaded_sg,
-            ont_graph=loaded_og,
-            options=validator_options_dict,
-        )
-        conforms, report_graph, report_text = validator.run()
-    except ValidationFailure as e:
-        conforms = False
-        report_graph = e
-        report_text = "Validation Failure - {}".format(e.message)
-    if do_check_dash_result and validator is not None:
-        passes = check_dash_result(validator, report_graph, loaded_sg or loaded_dg)
-        return passes, report_graph, report_text
-    do_serialize_report_graph = kwargs.pop('serialize_report_graph', False)
-    if do_serialize_report_graph and isinstance(report_graph, rdflib.Graph):
-        if not (isinstance(do_serialize_report_graph, str)):
-            do_serialize_report_graph = 'turtle'
-        report_graph = report_graph.serialize(None, encoding='utf-8', format=do_serialize_report_graph)
-    return conforms, report_graph, report_text
